# Add thread pooling for batch requests