--- a/adapters/repos/db/clusterintegrationtest/cluster_integration_test.go
+++ b/adapters/repos/db/clusterintegrationtest/cluster_integration_test.go
@@ -690,7 +690,8 @@ type node struct {
 }
 
 func (n *node) init(numberOfNodes int, dirName string, shardStateRaw []byte,
-	allNodes *[]*node) {
+	allNodes *[]*node,
+) {
 	localDir := path.Join(dirName, n.name)
 	logger, _ := test.NewNullLogger()
 
@@ -710,6 +711,7 @@ func (n *node) init(numberOfNodes int, dirName string, shardStateRaw []byte,
 		QueryMaximumResults:       10000,
 		DiskUseWarningPercentage:  config.DefaultDiskUseWarningPercentage,
 		DiskUseReadOnlyPercentage: config.DefaultDiskUseReadonlyPercentage,
+		MaxImportGoroutinesFactor: 1,
 	}, client, nodeResolver, nil)
 	n.schemaGetter = &fakeSchemaGetter{
 		shardState: shardState,
@@ -935,7 +937,8 @@ func exampleDataWithRefs(size int, refCount int, targetObjs []*models.Object) []
 }
 
 func bruteForceObjectsByQuery(objs []*models.Object,
-	query []float32) []*models.Object {
+	query []float32,
+) []*models.Object {
 	type distanceAndObj struct {
 		distance float32
 		obj      *models.Object
@@ -980,7 +983,8 @@ func normalize(v []float32) []float32 {
 
 func manuallyResolveRef(t *testing.T, obj *models.Object,
 	possibleTargets []*models.Object, localPropName,
-	referencedPropName string) []map[string]interface{} {
+	referencedPropName string,
+) []map[string]interface{} {
 	beacons := obj.Properties.(map[string]interface{})[localPropName].(models.MultipleRef)
 	out := make([]map[string]interface{}, len(beacons))