--- a/adapters/repos/db/init.go
+++ b/adapters/repos/db/init.go
@@ -55,6 +55,7 @@ func (d *DB) init(ctx context.Context) error {
 				DiskUseWarningPercentage:  d.config.DiskUseWarningPercentage,
 				DiskUseReadOnlyPercentage: d.config.DiskUseReadOnlyPercentage,
 				QueryMaximumResults:       d.config.QueryMaximumResults,
+				MaxImportGoroutinesFactor: d.config.MaxImportGoroutinesFactor,
 			}, d.schemaGetter.ShardingState(class.Class),
 				inverted.ConfigFromModel(invertedConfig),
 				class.VectorIndexConfig.(schema.VectorIndexConfig),