--- a/adapters/repos/db/migrator.go
+++ b/adapters/repos/db/migrator.go
@@ -38,6 +38,7 @@ func (m *Migrator) AddClass(ctx context.Context, class *models.Class,
 			DiskUseWarningPercentage:  m.db.config.DiskUseWarningPercentage,
 			DiskUseReadOnlyPercentage: m.db.config.DiskUseReadOnlyPercentage,
 			QueryMaximumResults:       m.db.config.QueryMaximumResults,
+			MaxImportGoroutinesFactor: m.db.config.MaxImportGoroutinesFactor,
 		},
 		shardState,
 		// no backward-compatibility check required, since newly added classes will