--- a/adapters/repos/db/repo.go
+++ b/adapters/repos/db/repo.go
@@ -72,6 +72,7 @@ type Config struct {
 	QueryMaximumResults       int64
 	DiskUseWarningPercentage  uint64
 	DiskUseReadOnlyPercentage uint64
+	MaxImportGoroutinesFactor float64
 }
 
 // GetIndex returns the index if it exists or nil if it doesn't