--- a/adapters/repos/db/shard.go
+++ b/adapters/repos/db/shard.go
@@ -41,6 +41,8 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
+const MaxImportGoroutinesFactorDefault = 1.5
+
 // Shard is the smallest completely-contained index unit. A shard manages
 // database files for all the objects it owns. How a shard is determined for a
 // target object (e.g. Murmur hash, etc.) is still open at this point
@@ -91,12 +93,16 @@ func NewShard(ctx context.Context, promMetrics *monitoring.PrometheusMetrics,
 	}
 
 	invertedIndexConfig := index.getInvertedIndexConfig()
-	val, ok := os.LookupEnv("MaxNumberGoroutinesFactor")
-	var MaxNumberGoroutinesFactor float64
+	val, ok := os.LookupEnv("MaxImportGoroutinesFactor")
+	var MaxImportGoroutinesFactor float64
 	if !ok {
-		MaxNumberGoroutinesFactor = 2
+		MaxImportGoroutinesFactor = MaxImportGoroutinesFactorDefault
 	} else {
-		MaxNumberGoroutinesFactor, _ = strconv.ParseFloat(val, 32)
+		MaxImportGoroutinesFactor, err = strconv.ParseFloat(val, 32)
+		// Use default value if we cannot parse input or it doesn't make any sense
+		if err != nil || MaxImportGoroutinesFactor <= 0 {
+			MaxImportGoroutinesFactor = MaxImportGoroutinesFactorDefault
+		}
 	}
 	s := &Shard{
 		index:            index,
@@ -112,7 +118,7 @@ func NewShard(ctx context.Context, promMetrics *monitoring.PrometheusMetrics,
 		randomSource:        rand,
 		diskScanState:       newDiskScanState(),
 		jobQueueCh:          make(chan job, 100000),
-		maxNumberGoroutines: int(math.Round(MaxNumberGoroutinesFactor * float64(runtime.GOMAXPROCS(0)))),
+		maxNumberGoroutines: int(math.Round(MaxImportGoroutinesFactor * float64(runtime.GOMAXPROCS(0)))),
 	}
 	defer s.metrics.ShardStartup(before)