--- a/test/benchmark/run_performance_tracker.sh
+++ b/test/benchmark/run_performance_tracker.sh
@@ -3,4 +3,4 @@
 # change to script directory
 cd "${0%/*}" || exit
 
-go run . -name "SIFT" -numberEntries 100000 -fail "-1"
+go run . -name "SIFT" -numberEntries 100000 -fail "-1" -numBatches "1"