--- a/usecases/config/config_handler.go
+++ b/usecases/config/config_handler.go
@@ -47,6 +47,7 @@ const (
 const (
 	DefaultDiskUseWarningPercentage  = uint64(80)
 	DefaultDiskUseReadonlyPercentage = uint64(90)
+	DefaultMaxImportGoroutinesFactor = float64(1.5)
 )
 
 // Flags are input options
@@ -56,23 +57,24 @@ type Flags struct {
 
 // Config outline of the config file
 type Config struct {
-	Name                    string         `json:"name" yaml:"name"`
-	Debug                   bool           `json:"debug" yaml:"debug"`
-	QueryDefaults           QueryDefaults  `json:"query_defaults" yaml:"query_defaults"`
-	QueryMaximumResults     int64          `json:"query_maximum_results" yaml:"query_maximum_results"`
-	Contextionary           Contextionary  `json:"contextionary" yaml:"contextionary"`
-	Authentication          Authentication `json:"authentication" yaml:"authentication"`
-	Authorization           Authorization  `json:"authorization" yaml:"authorization"`
-	Origin                  string         `json:"origin" yaml:"origin"`
-	Persistence             Persistence    `json:"persistence" yaml:"persistence"`
-	DefaultVectorizerModule string         `json:"default_vectorizer_module" yaml:"default_vectorizer_module"`
-	EnableModules           string         `json:"enable_modules" yaml:"enable_modules"`
-	ModulesPath             string         `json:"modules_path" yaml:"modules_path"`
-	AutoSchema              AutoSchema     `json:"auto_schema" yaml:"auto_schema"`
-	Cluster                 cluster.Config `json:"cluster" yaml:"cluster"`
-	Monitoring              Monitoring     `json:"monitoring" yaml:"monitoring"`
-	Profiling               Profiling      `json:"profiling" yaml:"profiling"`
-	DiskUse                 DiskUse        `json:"disk_use" yaml:"disk_use"`
+	Name                      string         `json:"name" yaml:"name"`
+	Debug                     bool           `json:"debug" yaml:"debug"`
+	QueryDefaults             QueryDefaults  `json:"query_defaults" yaml:"query_defaults"`
+	QueryMaximumResults       int64          `json:"query_maximum_results" yaml:"query_maximum_results"`
+	Contextionary             Contextionary  `json:"contextionary" yaml:"contextionary"`
+	Authentication            Authentication `json:"authentication" yaml:"authentication"`
+	Authorization             Authorization  `json:"authorization" yaml:"authorization"`
+	Origin                    string         `json:"origin" yaml:"origin"`
+	Persistence               Persistence    `json:"persistence" yaml:"persistence"`
+	DefaultVectorizerModule   string         `json:"default_vectorizer_module" yaml:"default_vectorizer_module"`
+	EnableModules             string         `json:"enable_modules" yaml:"enable_modules"`
+	ModulesPath               string         `json:"modules_path" yaml:"modules_path"`
+	AutoSchema                AutoSchema     `json:"auto_schema" yaml:"auto_schema"`
+	Cluster                   cluster.Config `json:"cluster" yaml:"cluster"`
+	Monitoring                Monitoring     `json:"monitoring" yaml:"monitoring"`
+	Profiling                 Profiling      `json:"profiling" yaml:"profiling"`
+	DiskUse                   DiskUse        `json:"disk_use" yaml:"disk_use"`
+	MaxImportGoroutinesFactor float64        `json:"max_import_goroutine_factor" yaml:"max_import_goroutine_factor"`
 }
 
 type moduleProvider interface {