This PR adds

- (optional) Batching to the benchmark script
- Add thread pooling for batch requests.
- Add monitoring of how many goroutines are started to grafana

Tests with different numbers of parallel batches with the benchmarking script via the remote machine. This PR is slightly slower, however it scales better with more parallel queries. Note that on my macbook the effect was stronger, eg it was slower with one batch and scaled much better with multiple batches.

| nr batches   |   master  |        This PR  |
| ------------- | ------------- | ------------- |
| 1  | 15056ms  | 15479ms |
| 2  | 27252ms  | 27488ms  |
| 5  | 46213ms  | 40499ms  |

There is currently an environment variable MaxNumberGoroutinesFactor that controls how many goroutines with respect to the number of cores are started. Not sure if this should be merged :)


Here is the profiling output (from my machine), I don't have a good idea why the PR is slower:

[CPU_PR](https://user-images.githubusercontent.com/5353192/184913852-85335014-3102-4c7b-91e7-4fdc331b3c9c.png)
[ALLOCS_PR](https://user-images.githubusercontent.com/5353192/184913885-e7b5835a-7055-4690-8984-5df99f2d3f75.png)
[CPU_master](https://user-images.githubusercontent.com/5353192/184914001-1dddaf0a-5ba5-4ce5-ae0e-b15329eefdc0.png)
[ALLOCS_master](https://user-images.githubusercontent.com/5353192/184914017-cfc37f4e-7012-411e-bd9a-417dac607e31.png)


[Chaos pipeline passed](https://github.com/semi-technologies/weaviate-chaos-engineering/runs/7874556610)